refactor: move handleBusyHello/handleIdleHello into wss.ts
Reviewer feedback: a separate file adds coupling and reduces cohesion
without meaningful benefit — the goal was to extract a long method,
not split a file. Move the exported functions back into wss.ts as
module-level exports and delete hello-handlers.ts.
