+++ CLAUDE.md
@@ -86,5 +86,5 @@ See **`docs/type-system.md`** for the full design. In brief: one server model cl
 - In unit tests, use `setupInMemoryTasks()` from `tests/helpers/task.ts` instead of `initDb()` — it mocks the `Task` layer with an in-memory map. `WebhookEvent.log()` and `ForemanMessage.log()` return `Promise<void>` (guarded with `if (!db) return Promise.resolve()` inside a try-catch, so they silently no-op when Supabase is not initialized). Production callers use `void` to discard the promise; tests can `await` them for deterministic setup.
 - Wire protocol types (foreman↔worker) live in `src/wire.ts` and are imported as `import * as Wire from "./wire.js"`. The DB model class for foreman messages is `ForemanMessage` (in `src/foreman/models/foreman-message.ts`), distinct from the wire type `Wire.ForemanMessage`.
 - In agent tests that need slash commands, create a `CommandRegistry` instance directly (`new CommandRegistry()`) and populate it via `registerTestCommands()` from `tests/helpers.ts` (which returns the registry). Pass the registry to functions like `dispatchInput`, `parseSlashCommand`, etc.; call `registry.listCommandNames()` and `registry.listCommands()` directly on the registry (with injectable `listDir`/`readFile` args for testing). When calling `registerWorkspaceCommands` or `registerWorkerCommands` in tests, pass a pre-scoped registry (e.g. `registry.scoped("workspace")`) — scoping is always the caller's responsibility.
-- In foreman tests that call `createForemanWss`, call `Worker._reset()` in `beforeEach` to clear the module-level worker registry between tests. `createForemanWss` takes no `registry` parameter — `Worker` is imported directly by `wss.ts` and `event-router.ts`.
+- In foreman tests that call `createForemanWss`, call `Worker._reset()` in `beforeEach` to clear the module-level worker registry between tests. `createForemanWss` takes no `registry` parameter — `Worker` is imported directly by `wss.ts` and `event-router.ts`. To unit-test the busy/idle reconnection logic without a WebSocket server, import `handleBusyHello` / `handleIdleHello` from `wss.ts` and call them directly with mocked deps.
 - In browser tests, the server is shared across all tests — use unique issue numbers per test rather than resetting server state.
